refactor(seed): split the manifest transport out, and close the review's remaining findings
manifest.py carried the document and the way it leaves a pod; the transport
moves to manifest_sentinel.py, typed against Mapping so nothing imports back.
The sentinel tests follow it into their own file. 522 -> 421 and 433 -> 317
lines, with two focused modules beside them.

Tests that passed on a mutation now fail on it: the growth surname pool is
pinned to a written-out golden head rather than compared to its own output;
every dev credential and every credential-bearing key word is exercised and
cross-checked against the module's own constants; the published field set is
pinned, because the version gate only fires on a bump and not on a silent
change.

seed-stand.sh: --dry-run no longer requires uv, --manifest-out is refused
where it would be a no-op, the ConfigMap is driven by whether a step produces
a manifest rather than by its name, a Job that emitted none says so before
the local document is substituted, and a partial read of the rendered Job's
environment now dies as loudly as a total one.

The contract test's module shim unwinds itself on a partial load; the
full-refresh fixture patches the module's subprocess handle instead of
stdlib's. Docs corrected where they had drifted: both workflows now take the
manifest from the seeder, and the ceiling names the two bounds that hold it.
