+++ deploy/gitops/Makefile
@@ -416,7 +416,6 @@ keycloak-realm: kube-ctx values-present
 	echo "$(C_GRN)→$(C_RST) generating Keycloak realm (dev-lead $$DEV_EMAIL; redirects $$REDIRECT, $$LB_REDIRECT)"; \
 	TENANT_ID="$$(yq -r '.global.tenantDefaultId // ""' $(VALUES))"; \
 	[ -n "$$TENANT_ID" ] || { echo "$(C_RED)global.tenantDefaultId is required to generate the realm$(C_RST)"; exit 1; }; \
-	: "$${SEED_ORG_HEADCOUNT:=}"; \
 	[ -z "$$SEED_ORG_HEADCOUNT" ] || echo "$(C_GRN)→$(C_RST) realm sized for SEED_ORG_HEADCOUNT=$$SEED_ORG_HEADCOUNT people"; \
 	SEED_ORG_HEADCOUNT="$$SEED_ORG_HEADCOUNT" TENANT_DEFAULT_ID="$$TENANT_ID" uv run --project "$$SEED_DIR" insight-seed-realm --dev-email "$$DEV_EMAIL" \
 	  --authenticator-redirect "$$REDIRECT" \
