+++ deploy/gitops/scripts/recreate-test-stand.sh
@@ -330,7 +330,7 @@ fi
 if [ "$KEEP_DATABASES" != "1" ]; then
   hdr "database wipe"
   # Unowned by helm, so `uninstall` leaves it describing rows this wipe drops.
-  "${KUBECTL[@]}" -n "$NAMESPACE" delete configmap seed-stand-manifest --ignore-not-found >/dev/null 2>&1 || true
+  "${KUBECTL[@]}" -n "$NAMESPACE" delete configmap seed-stand-manifest --ignore-not-found >/dev/null
   WIPE_JOB="recreate-wipe-$(date -u +%Y%m%d%H%M%S)"
   wipe_manifest() {
     cat <<YAML
